requirements.txt
streamlit>=1.25.0
pandas>=2.0.0
numpy>=1.23.0
pillow>=9.0.0
matplotlib>=3.7.0
plotly>=5.14.0
